move all the functions related to array plot to arrayplot.py :D
